El0ps.AbstractDatafit — Type

AbstractDatafit

Abstract type for the function F(y,.) in Problem.

El0ps.bind_model! — Method

bind_model!(F::AbstractDatafit, y::Vector, model::JuMP.Model)

Formulate the function F(y,.) in model using an epigraph method. See initialize_model for more details.

El0ps.conjugate — Method

conjugate(F::AbstractDatafit, y::Vector, w::Vector)

Fenchel conjugate of F(y,.).

El0ps.dual_scale! — Method

dual_scale!(F::AbstractDatafit, y::Vector, u::Vector)

In-place transform a vector u into a feasible one for the conjugate of F(y,.).

El0ps.gradient — Method

gradient(F::AbstractDatafit, y::Vector, w::Vector)

Gradient of F(y,.).

El0ps.lipschitz_constant — Method

lipschitz_constant(F::AbstractDatafit, y::Vector)

Lischitz constant of F(y,.).

El0ps.value — Method

value(F::AbstractDatafit, y::Vector, w::Vector)

Value of F(y,.).

El0ps.LeastSquares — Type

LeastSquares

Least-squares function F(y,w) = (1/m) * ||y-w||_2^2 where m=size(y).

El0ps.Logistic — Type

Logistic

Logistic function F(y,w) = (1/m) * sum(log(1 + exp(- y ⊙ w))) where m=size(y), where ⊙ denotes the Hadamard product and where the log and the exp functions are taken component-wisely.