El0ps.AbstractPenalty — Type

AbstractPenalty

Abstract type for the function G(.) in Problem.

El0ps.bind_model! — Method

bind_model!(G::AbstractPenalty, model::JuMP.Model)

Formulate the function G(.) in model using an epigraph method. See initialize_model for more details.

El0ps.conjugate_1d — Method

conjugate_1d(G::AbstractPenalty, v::Float64)

Conjugate of the function G(.) evaluated at some scalar v.

El0ps.dual_scale! — Method

dual_scale!(G::AbstractPenalty, A::Matrix, u::Vector, λ::Float64)

In-place transform any vector u into a feasible one for the conjugate of the function G(.) evaluated at A'u. Returns the value of A'u.

El0ps.params_to_dict — Method

params_to_dict(G::AbstractPenalty)

Returns a dictionary with the parameters names of the function G(.) and their value. The dictionary is empty is G(.) has no parameters.

El0ps.prox_1d — Method

prox_1d(G::AbstractPenalty, x::Float64, η::Float64)

Proximal mapping of the function ηG(.) evaluated at some scalar x.

El0ps.value_1d — Method

value_1d(G::AbstractPenalty, x::Float64)

Value of G(.) evaluated at some scalar x.

El0ps.Bigm — Type

Bigm

Big-M function G(x) = Ind(||x||_Inf <= M).

Arguments

M::Float64 : Big-M value.