Bounding step

El0ps.AbstractBoundingSolver — Type

AbstractBoundingSolver

Solver for the lower and upper bounding steps in the BnbSolver.

El0ps.CD — Type

CD

Coordinate Descent solver for the lower and upper bounding steps in the BnbSolver.

Arguments

tolgap::Float64 : Absolute tolearance on the duality gap.

maxiter::Int : Maximum number of itrations.

El0ps.CDAS — Type

CDAS

Coordinate Descent with Active-Set strategy solver for the lower and upper bounding steps in the BnbSolver.

Arguments

reltol::Float64 : Relative objective improvment tolearance for lower bouning.

maxiter_cd::Int : Maximum number of cooridinate descent updates.

maxiter_as::Int : Maximum number of active-set updates.