Data generation

El0ps.synthetic_data_classification — Method

synthetic_data_classification(
    k::Int,
    m::Int,
    n::Int,
    ρ::Float64,
    σ::Float64;
    normalize::Bool=false
)

Generate a synthetic classification instance of a Problem with y ≈ Bernoulli(p) where y[j] = 1 with probability p[j] = 1 / (1 + exp(-σ[Ax]_j)) and y[j] = -1 otherwise.

Arguments

k::Int : Number of non-zeros in x.

m::Int : Number of rows in A.

n::Int : Number of columns in A.

ρ::Float64 : Correlation between the columns in A.

σ::Float64 : Probability parameter.

normalize::Bool=false : Whether to normalize the columns in A.

El0ps.synthetic_data_regression — Method

synthetic_data_regression(
    k::Int,
    m::Int,
    n::Int,
    ρ::Float64,
    σ::Float64;
    normalize::Bool=false
)

Generate a synthetic regression instance of a Problem with y = Ax + ϵ where x is a k-sparse vector.

Arguments

k::Int : Number of non-zeros in x.

m::Int : Number of rows in A.

n::Int : Number of columns in A.

ρ::Float64 : Correlation between the columns in A.

σ::Float64 : SNR ratio of ϵ with respect to Ax.

normalize::Bool=false : Whether to normalize the columns in A.