Datafits

El0ps.AbstractDatafit — Type

AbstractDatafit

Abstract type for the function F in a Problem.

El0ps.bind_model! — Method

bind_model!(F::AbstractDatafit, y::Vector, model::JuMP.Model)

Formulate the function F in model. See initialize_model for more details about the argument model.

El0ps.conjugate — Method

conjugate(F::AbstractDatafit, y::Vector, w::Vector)

Fenchel conjugate of F with respect to its second argument.

El0ps.dual_scale! — Method

dual_scale!(F::AbstractDatafit, y::Vector, u::Vector)

In-place transforms a vector u into a feasible one for the conjugate of F.

El0ps.gradient — Method

gradient(F::AbstractDatafit, y::Vector, w::Vector)

Gradient of F with respect to its second argument.

El0ps.lipschitz_constant — Method

lipschitz_constant(F::AbstractDatafit, y::Vector)

Lischitz constant of the gradient of F with respect to its second argument.

El0ps.value — Method

value(F::AbstractDatafit, y::Vector, w::Vector)

Value of F.

El0ps.LeastSquares — Type

LeastSquares <: AbstractDatafit

Least-squares function

$F(\mathbf{y},\mathbf{w}) = \tfrac{1}{m} \|\mathbf{y}-\mathbf{w}\|_2^2$

where m = length(y).

El0ps.Logistic — Type

Logistic <: AbstractDatafit

Logistic function

$F(\mathbf{y},\mathbf{w}) = \tfrac{1}{m} \sum(\log(\mathbf{1} + \exp(- \mathbf{y} \odot \mathbf{w})))$

where m = length(y), where $\odot$ denotes the Hadamard product and where the log and the exp functions are taken component-wisely.