Regularization path

El0ps.Path — Type

Path

Regularization path of a Problem, i.e., solutions for different values of λ.

El0ps.PathOptions — Type

PathOptions

Options for a Path. The path is computed over a logarithmically-spaced grid λ ∈ [λratio_max, λratio_min] * λmax of λratio_num different values. The value of λmax is computed using compute_λmax.

Arguments

λratio_max::Float64 : Maximum value of λ/λmax.

λratio_min::Float64 : Minimum value of λ/λmax.

λratio_num::Int : Number of values of λ in the regularization path.

max_support_size::Int : Stop the path fitting when a solution with support

size max_support_size is recovered.

stop_if_unsolved::Bool : If true, stop the path fitting if the

Problem at some λ is unsolved.

compute_cv::Bool : If true, compute the cross-validation error over

nb_folds folds for each solution obtained in the Path.

nb_folds::Int : Number of cross-validation folds.

verbosity::Bool : Whether to displays outputs.

El0ps.fit_path — Method

fit_path(
    solver::AbstractSolver,
    f::AbstractDatafit,
    h::AbstractPenalty,
    A::Matrix,
    kwargs...
)

Fit a regularization Path for a Problem. Additional keyword arguments in kwargs are passed to a PathOptions instance.