Penalties

El0ps.AbstractPenalty — Type

AbstractPenalty

Abstract type for the function G in a Problem. This function is supposed to split. All its associated methods are therefore implemented with one-dimensional inputs.

El0ps.bind_model! — Method

bind_model!(G::AbstractPenalty, model::JuMP.Model)

Formulate the function G in model. See initialize_model for more details about the argument model.

El0ps.conjugate — Method

conjugate(G::AbstractPenalty, v::Vector)

Value of the conjugate of G evaluated with a vector v as input.

El0ps.conjugate_1d — Method

conjugate_1d(G::AbstractPenalty, v::Float64)

Value of the conjugate of G.

El0ps.conjugate_vectorized — Method

conjugate_vectorized(G::AbstractPenalty, v::Vector)

Coordinate-wise value of the conjugate of G when evaluated with a vector v as input.

El0ps.dual_scale! — Method

dual_scale!(G::AbstractPenalty, A::Matrix, u::Vector, λ::Float64)

In-place transforms some vector u into a feasible one for the conjugate of G when evaluated at A'u. Returns the value of A'u.

El0ps.params_to_dict — Method

params_to_dict(G::AbstractPenalty)

Returns a dictionary with the parameters name and value of the function G.

El0ps.prox — Method

prox(G::AbstractPenalty, v::Vector, η::Float64)

Proximal mapping of the function ηG evaluated with a vector v as input.

El0ps.prox_1d — Method

prox_1d(G::AbstractPenalty, x::Float64, η::Float64)

Proximal mapping of the function ηG.

El0ps.value — Method

value(G::AbstractPenalty, v::Vector)

Value of G evaluated with a vector v as input.

El0ps.value_1d — Method

value_1d(G::AbstractPenalty, x::Float64)

Value of G.

El0ps.Bigm — Type

Bigm <: AbstractPenalty

Convex indicator of a Big-M constraint

$G(\mathbf{x}) = \mathrm{Ind}(\|\mathbf{x}\|_{\infty} \leq M)$

where M > 0.

Arguments

M::Float64 : Big-M value.