Problem

El0ps.Problem — Type

Problem

Problem of the form

$\min F(\mathbf{y},\mathbf{Ax}) + \lambda (\|x\|_0 + G(\mathbf{x}))$

where $\mathbf{A} \in \mathbf{R}^{m \times n}$, $\mathbf{y} \in \mathbf{R}^{m}$ and $\lambda > 0$.

El0ps.compute_λmax — Method

compute_λmax(F::AbstractDatafit, G::AbstractPenalty, A::Matrix, y::Vector)

Return a value of λ such that 0 is a solution of a Problem.

El0ps.objective — Method

objective(problem::Problem, x::Vector, Ax::Vector)

Returns the value of the objective of a Problem when Ax is already computed.

El0ps.objective — Method

objective(problem::Problem, x::Vector)

Returns the value of the objective of a Problem.