Solvers

El0ps.BnbNodeStatus — Type

NodeStatus

Status of a BnbNode.

OPEN : The node has not been treated yet.

PRUNED : The node has been pruned

SOLVED : The node has been treated and not pruned.

PERFECT : The node has a perfect relaxation.

El0ps.BnbNodeType — Type

NodeType

Tyep of a BnbNode.

ROOT : Root node.

ZERO : Created by setting an entry to zero.

ONE : Created by setting an entry to non-zero.

El0ps.BnbOptions — Type

BnbOptions

Options of a BnbSolver.

El0ps.BnbOptions — Method

BnbOptions(;
    lb_solver::AbstractBoundingSolver   = CDAS(LOWER_BOUNDING),
    ub_solver::AbstractBoundingSolver   = CDAS(UPPER_BOUNDING),
    exploration::ExplorationStrategy    = DFS,
    depthswitch::Int                    = 10,
    branching::BranchingStrategy        = LARGEST,
    maxtime::Float64                    = 60.,
    maxnode::Int                        = typemax(Int),
    tolgap::Float64                     = 1e-8,
    tolint::Float64                     = 1e-8,
    tolprune::Float64                   = 0.,
    dualpruning::Bool                   = false,
    l0screening::Bool                   = false,
    l1screening::Bool                   = false,
    verbosity::Bool                     = false,
    showevery::Int                      = 1,
    keeptrace::Bool                     = false,
)

Instantiate a BnbOptions.

Keywords:

lb_solver::AbstractBoundingSolver : Solver for the lower-bounding step.

ub_solver::AbstractBoundingSolver : Solver for the upper-bounding step.

exploration::ExplorationStrategy : Tree exploration strategy.

depthswitch::Int : The depth where the MIXED ExplorationStrategy

is switched.

branching::BranchingStrategy : Tree branching strategy.

maxtime::Float64 : Maximum solution time in seconds.

maxnode::Int : Maximum number of nodes.

tolgap::Float64 : Relative MIP gap tolerance.

tolint::Float64 : Integer tolerance, i.e., x = 0 when |x| < tolint.

tolprune::Float64 : Prune a node in the bnb tree when `bnb.ub +

tolprune < node.lb`.

dualpruning::Bool : Toogle the dual-pruning acceleration.

l0screening::Bool : Toogle the L0-screening acceleration.

l1screening::Bool : Toogle the L1-screening acceleration.

verbosity::Bool : Toogle verbosity.

showevery::Int : Displays logs every showevery nodes explored.

keeptrace::Bool : Whether to fill the BnbTrace or not.

El0ps.BnbResult — Type

BnbResult

Result of a BnbSolver.

El0ps.BnbSolver — Type

BnbSolver

Branch-and-Bound solver for a Problem.

El0ps.BnbSolver — Method

BnbSolver(; kwargs...)

Instantiate a BnbSolver. Keywords are passed to BnbOptions.

El0ps.BnbTrace — Type

BnbTrace

Trace of a BnbSolver.

El0ps.BranchingStrategy — Type

BranchingStrategy

Branching strategy of a BnbSolver.

LARGEST : Select the largest index in absolute value in the relaxation

solution.

RESIDUAL : Select the largest index in absolute value in the dual residual.

El0ps.ExplorationStrategy — Type

ExplorationStrategy

Exploration strategy of a BnbSolver.

BFS : Breadth-First Search

DFS : Depth-First Search

MIXED : Mixed exploration strategy where the k-top layers of nodes are

explored in a DFS fashion and where the below ones are explored in a BFS fashion. The parameter k is specified in BnbOptions.

El0ps.optimize — Method

optimize(
    solver::BnbSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
    S0::Vector{Int}=Vector{Int}(),
    S1::Vector{Int}=Vector{Int}(),
)

Optimize a Problem with a BnbSolver. The argument x0 is used as a warm start. The arguments S0 and S1 can be used to impose zero and non-zero constraints directly in the root node. They must match x0.