Solvers

El0ps.AbstractResult — Type

AbstractResult

Abstract type to store results of an AbstractSolver.

El0ps.AbstractSolver — Type

AbstractSolver

Abstract type for solver tailored to the problem min F(y,Ax) + λ * (||x||_0 + G(x)).

El0ps.optimize — Method

optimize(
    solver::AbstractSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
)

Optimize Problem with the solver solver using the warm start x0.

El0ps.BnbOptions — Type

BnbOptions

Options of BnbSolver.

Arguments

lb_solver::AbstractBoundingSolver = CoordinateDescent(tolgap=1e-4, maxiter=10_000) : Solver for the lower-boudning step.

ub_solver::AbstractBoundingSolver = CoordinateDescent(tolgap=1e-8, maxiter=10_000) : Solver for the upper-boudning step.

exploration::ExplorationStrategy = DFS : Tree exploration strategy.

branching::BranchingStrategy = LARGEST : Tree branching strategy.

maxtime::Float64 = 60. : Maximum solution time in seconds.

maxnode::Int = typemax(Int) : Maximum number of nodes.

tolgap::Float64 = 0. : Relative MIP gap tolerance.

tolperf::Float64 = 0. : When node.ub - node.lb < tolperf, consider that

the relaxation of node is perfect.

tolint::Float64 = 0. : Integer tolerance, ie x = 0 when |x| < tolint.

tolprune::Float64 = 0. : Prune a node when ub + tolprune < node.lb.

dualpruning::Bool = false : Toogle the dual pruning acceleration.

l0screening::Bool = false : Toogle the L0-screening acceleration.

l1screening::Bool = false : Toogle the L1-screening acceleration.

verbosity::Bool = false : Toogle the verbosity.

showevery::Int = 1 : Displays logs every showevery nodes.

keeptrace::Bool = true : Whether to keep the solver trace or not.

El0ps.BnbResult — Type

BnbResult

Result of BnbSolver.

El0ps.BnbSolver — Type

BnbSolver

Branch-and-Bound algorithm to solve Problem. Keyword arguments are passed to BnbOptions.

El0ps.BnbTrace — Type

BnbTrace

Trace of BnbSolver.

El0ps.BranchingStrategy — Type

BranchingStrategy

Branching strategy of BnbSolver :

LARGEST : Select the largest index in absolute value in the relaxation

solution.

RESIDUAL : Select the largest index in absolute value in the vector A'u

where u is the dual relaxation solution.

El0ps.ExplorationStrategy — Type

ExplorationStrategy

Exploration strategy of BnbSolver :

BFS : Breadth-First Search

DFS : Depth-First Search

El0ps.optimize — Method

optimize(
    solver::BnbSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
)

Optimize Problem with a BnbSolver using the warm start x0.

El0ps.DirectResult — Type

DirectResult

Result of DirectSolver.

El0ps.DirectSolver — Type

DirectSolver

Direct solver for a Problem.

Arguments

optimizer : An optimizer callable with JuMP such as GLPK.Optimizer.

options::Dict : Options name and value to pass to optimizer with the

method set_optimizer_attribute.

El0ps.initialize_model — Method

initialize_model(problem::Problem, solver::DirectSolver, x0::Vector)

Initialize the JuMP model

```
min Fcost + λ Ωcost
 st Ax = w
    x ∈ R^n
    z ∈ {0,1}^n
    w ∈ R^m
    Fcost ∈ R
    Gcost ∈ R
```

and set the initial x to x0. To complete the model, one has to formulate the expressions of the function F and G in the problem using bind_model!.

El0ps.optimize — Method

optimize(
    solver::DirectSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
)

Optimize Problem with a DirectSolver using the warm start x0.