Solvers

El0ps.AbstractResult — Type

AbstractResult

Abstract type to store results of an AbstractSolver.

El0ps.AbstractSolver — Type

AbstractSolver

Abstract type for a solver tailored to a Problem.

El0ps.optimize — Method

optimize(
    solver::AbstractSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
)

Optimize a Problem with an AbstractSolver. The argument x0 is used as a warm start.

El0ps.BnbOptions — Type

BnbOptions

Options of a BnbSolver.

Arguments

lb_solver::AbstractBoundingSolver : Solver for the lower-bounding step.

ub_solver::AbstractBoundingSolver : Solver for the upper-bounding step.

exploration::ExplorationStrategy : Tree exploration strategy.

branching::BranchingStrategy : Tree branching strategy.

maxtime::Float64 : Maximum solution time in seconds.

maxnode::Int : Maximum number of nodes.

tolgap::Float64 : Relative MIP gap tolerance.

tolint::Float64 : Integer tolerance, i.e., x = 0 when |x| < tolint.

tolprune::Float64 : Prune a node in the bnb tree when bnb.ub + tolprune < node.lb.

dualpruning::Bool : Toogle the dual-pruning acceleration.

l0screening::Bool : Toogle the L0-screening acceleration.

l1screening::Bool : Toogle the L1-screening acceleration.

verbosity::Bool : Toogle verbosity.

showevery::Int : Displays logs every showevery nodes explored.

keeptrace::Bool : Whether to fill the BnbTrace or not.

El0ps.BnbResult — Type

BnbResult

Result of a BnbSolver.

El0ps.BnbSolver — Type

BnbSolver <: AbstractSolver

Branch-and-Bound algorithm to solve a Problem. Additional keyword arguments in kwargs are passed to a BnbOptions instance.

El0ps.BnbTrace — Type

BnbTrace

Trace of a BnbSolver.

El0ps.BranchingStrategy — Type

BranchingStrategy

Branching strategy of a BnbSolver.

LARGEST : Select the largest index in absolute value in the relaxed solution.

El0ps.ExplorationStrategy — Type

ExplorationStrategy

Exploration strategy of a BnbSolver.

BFS : Breadth-First Search

DFS : Depth-First Search

El0ps.optimize — Method

optimize(
    solver::BnbSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
    S0::Vector{Int}=Vector{Int}(),
    S1::Vector{Int}=Vector{Int}(),
)

Optimize a Problem with a BnbSolver. The argument x0 is used as a warm start. The arguments S0 and S1 can be used to impose zero and non-zero constraints directly in the root node. They must match x0.

El0ps.DirectResult — Type

DirectResult <: AbstractResult

Result of a DirectSolver.

El0ps.DirectSolver — Type

DirectSolver <: AbstractSolver

Formulate a Problem as a Mixed-Integer-Program and solve it using the specified optimizer.

Arguments

optimizer : The MOI.AbstractOptimizer of the specified optimizer (example

: GLPK.Optimizer).

options::Dict : Options name and value to pass to the

optimizer with the method set_optimizer_attribute from MathOptInterface.

El0ps.initialize_model — Method

initialize_model(problem::Problem, solver::DirectSolver, x0::Vector)

Initialize the JuMP model

min Fcost + λ Ωcost
st  Ax = w
    x ∈ R^n
    z ∈ {0,1}^n
    w ∈ R^m
    Fcost ∈ R
    Gcost ∈ R

and set the initial value of x to x0. To complete the model, one has to construct the epigraph formulations of the functions F and G in the Problem using bind_model! and the scalar values Fcost and Gcost that represent the epigraph value.

El0ps.optimize — Method

optimize(
    solver::DirectSolver,
    problem::Problem;
    x0::Union{Vector,Nothing}=nothing,
)

Optimize a Problem with a DirectSolver. The argument x0 is used as a warm start. The arguments S0 and S1 can be used to impose zero and non-zero constraints directly in the root node. They must match x0.